Refactor Sockets to Remove Redundant Code
